test: Move test macros in its own helper module
